renaming of modules
