+++ src/probnum/filtsmooth/gaussfiltsmooth/__init__.py
@@ -1,9 +1,9 @@
 """Gaussian filtering and smoothing."""
 
-from .extendedkalman import ContinuousEKFComponent, DiscreteEKFComponent, EKFComponent
-from .iterated_component import IteratedDiscreteComponent
-from .kalman import Kalman
-from .kalmanposterior import FilteringPosterior, KalmanPosterior, SmoothingPosterior
-from .stoppingcriterion import StoppingCriterion
-from .unscentedkalman import ContinuousUKFComponent, DiscreteUKFComponent, UKFComponent
-from .unscentedtransform import UnscentedTransform
+from ._extendedkalman import ContinuousEKFComponent, DiscreteEKFComponent, EKFComponent
+from ._iterated_component import IteratedDiscreteComponent
+from ._kalman import Kalman
+from ._kalmanposterior import FilteringPosterior, KalmanPosterior, SmoothingPosterior
+from ._stoppingcriterion import StoppingCriterion
+from ._unscentedkalman import ContinuousUKFComponent, DiscreteUKFComponent, UKFComponent
+from ._unscentedtransform import UnscentedTransform
