+++ src/probnum/filtsmooth/gaussfiltsmooth/_kalmanposterior.py
@@ -13,8 +13,8 @@ from probnum import _randomvariablelist, randvars, statespace, utils
 from probnum.type import FloatArgType, RandomStateArgType, ShapeArgType
 
 from ..timeseriesposterior import DenseOutputLocationArgType, TimeSeriesPosterior
-from .extendedkalman import ContinuousEKFComponent, DiscreteEKFComponent
-from .unscentedkalman import ContinuousUKFComponent, DiscreteUKFComponent
+from ._extendedkalman import ContinuousEKFComponent, DiscreteEKFComponent
+from ._unscentedkalman import ContinuousUKFComponent, DiscreteUKFComponent
 
 GaussMarkovPriorTransitionArgType = Union[
     statespace.DiscreteLinearGaussian,
