+++ src/probnum/filtsmooth/gaussfiltsmooth/_stoppingcriterion.py
