+++ tests/test_filtsmooth/test_gaussfiltsmooth/test_unscentedtransform.py
@@ -2,15 +2,14 @@ import unittest
 
 import numpy as np
 
-from probnum import randvars
-from probnum.filtsmooth.gaussfiltsmooth import unscentedtransform
+from probnum import filtsmooth, randvars
 
 
 class TestUnscentedTransform(unittest.TestCase):
     def setUp(self):
         self.ndim = np.random.randint(1, 33)  # 1 < random int < 33
         alpha, beta, kappa = np.random.rand(3)
-        self.ut = unscentedtransform.UnscentedTransform(self.ndim, alpha, beta, kappa)
+        self.ut = filtsmooth.UnscentedTransform(self.ndim, alpha, beta, kappa)
         self.mean = np.random.rand(self.ndim)
         cvr = np.random.rand(self.ndim, self.ndim)
         self.covar = cvr @ cvr.T
